+++ src/points/index.ts
@@ -1,0 +1,6 @@
+export * from './instruction_builders';
+export * from './pda_finders';
+export * from './types/points_accounts';
+export * from './types/points_program';
+export * from './utils';
+
