+++ src/points/instruction_builders/index.ts
@@ -1,0 +1,10 @@
+export * from './BaseParams'
+export * from './createUserPointAccountIx'
+export * from './createUserPointAccountWithLicenseIx'
+export * from './deregisterPointCategoryAccountIx'
+export * from './deregisterPointModifierIx'
+export * from './initializeDomainIx'
+export * from './modifyPointsIx'
+export * from './registerPointCategoryAccountIx'
+export * from './registerPointModifierIx'
+export * from './updatePointCategoryAccountIx'
