+++ src/points/pda_finders/seeds.ts
@@ -1,0 +1,4 @@
+export const DOMAIN_ACCOUNT_SEED = Buffer.from('"PointsDomain"');
+export const POINT_CATEGORY_ACCOUNT_SEED = Buffer.from('"PointCategoryAccount"');
+export const POINT_MODIFIER_SEED = Buffer.from('"PointModifier"');
+export const USER_POINTS_ACCOUNT_SEED = Buffer.from('"UserPointsAccount"');
