+++ src/points/utils/index.ts
@@ -1,0 +1,2 @@
+export * from './getPointsProgram'
+export * from './getters'
